refactor(rewards): extract RewardsDiscountBadge into its own component
Move the generic discount badge out of RewardsVipBadge so each file matches its exported component and bridge imports can target the shared badge directly.
