+++ ui/pages/bridge/prepare/bridge-discount-fee-message.tsx
@@ -10,10 +10,8 @@ import {
 import { useI18nContext } from '../../../hooks/useI18nContext';
 import { JustifyContent } from '../../../helpers/constants/design-system';
 import { Row } from '../layout';
-import {
-  RewardsDiscountBadge,
-  RewardsVipBadge,
-} from '../../../components/app/rewards/RewardsVipBadge';
+import { RewardsDiscountBadge } from '../../../components/app/rewards/RewardsDiscountBadge';
+import { RewardsVipBadge } from '../../../components/app/rewards/RewardsVipBadge';
 import { readMmFee } from '../utils/quote';
 
 export const BridgeDiscountFeeMessage = () => {
